(-4 example of google viz library)
Summary: demonstrate that we are able to use the first (simplest?)
example of google's visualization library (a piechart with hard-coded
data).

this demo works on our browser provided that we have an internet
connection.

NOTE the HTML and javascirpt committed here is *exactly* as-is
on their website at the time. none of this is the committer's
original work.

And:
  - add “TODO.stack” file
  - add documentation about same