( 12.2  birth of “node table” (4))
Discussion: in order to preview the changes we make in markdown before
we push them up to github, we added a utility script to generate html
from markdown. (we wrote it in ruby because the lib github uses is in
ruby.)

this sent us on a mind-storm thinking of a github-themed static
site generator, but later for that. (they have that, don’t they?)

also (or in service of, we can’t remember which) we’re experimenting
with what we describe in [#002] under ‘experiments’..