( 10.3  server with `select` example (refactored)..)
Discussion: this commit is a refactor of the work in the previous
commit. it adds no new features or behavior.

The server:

a simple socket server that listens on a port. when it receives ONE
connection (a client) it loops FOREVER (just for that one connection -
it will never see any other connections).

when reading from the client, it BLOCKS until data is available. when
data is available it reads ONLY 8 bytes of it. it turns this data (as
a string) to an uppercase string and sends it back. (it sends at most
8 bytes back at one time)

The client:

the client’s main job is simply to accept input from the user (any
string) and send it to the server. with the simplest mechanics we
would block on reading from the server and also block on reading
from the user. under the easiest of circumstances, this would work.

however here we have chunked our messages from the server into chunks
8 bytes long. when we send the client a message (string) longer than
8 bytes, the response from the server will be broken up over multiple
messages.

our current solution so we don’t block indefinitely when reading from
the server is to set a timeout (using `select`). (this timeout can in
fact be changed to be shorter or longer, from the “interface”).

this is an imperfect solution, however. a timeout too slow effects user
experience, and a timeout too quick doesn’t give the server time to
build and send its final message(s). we’ll explore better solutions to
this problem in the coming commits.