(+grep-dump.09.14 refactor: rewrite server startup (35))
Discussion: the server startup code had been “exploded” partly so we
could figure out why it wasn’t closing its open sockets on shutdown
(this issued a warning, and it seemed like bad style), partly so we
could get some ideas for how to hook in to things like server shutdown.

for a moment we thought that this “exploded” code was extraneous and
we moved towards closing it back up (and almost committed this) BUT
THEN we discovered:

  - we really want to “hook in” to the existing mechanism in order to
    carry out our own cleanup. it’s plainly an event-based thing
    (waiting for a keyboard interrupt) and it would be silly to try
    to pile-on more listener callbacks to that event through any means
    other than using the existing rigging.

  - we discovered that the reloader is kind of a bad implementation.
    (more on this later)

  - we discovered that our “jobser” won’t work with the reloader
    anyway.

As such, we wrote all the hacky “explosion” code BACK IN, BY HAND,
and we think we have improved it somewhat.

As a detour along the way, we also discovered that the reloader *had*
been working an *AWFUL* way (asking for a stat on *every* file *every*
N seconds???), so we added watchdog to our pipfile towards this:

  - added `watchdog` to the pipfile because yikes without it is yikes

  - (adding werkzeug to the pipfile did nothing but change one hash)

(0.015s)