(+sakin-agac.09.13 abstract/expose argparse fix (172))
Discussion:
while writing the CLI interface for the “sync” script we learned that
yes indeed we want better abstraction and exposure from what currently
comprises the bulk of the `game_server` sub-project..

at a minimum we wanted to be able to use the “fixing” of argparse
we had done there.

furthermore there will perhaps be a desire to further expose work
done in `game_server` (see the notes at the new [mam]).

  - there is a drop in the number of tests (down to 172 from 180)
    only because we aren't bothering running the API server tests.