(+sakin-agac.09.46 (2   ) behavior points: 3 more error cases (116))
Discussion: originally we were supposed to test the stack integration of
these, but they just throw happenstance exceptions for now. at least the
cases are named and covered.)

Also: “context manager” replaces “stream” as the dominant paradigm
for traversing a collection (at least, it does so locally for now).
(we don't love the nested context managers closing each other thing
and are realizing now we shoud probably not need to do that. it's
a violation of encapsulation (implementation detail) to "know" that
each nested context manager is one. but anyway this current state is
better that previous.)