(+sakin-agac.09.48 (3 -5) cover script: khong JSON object stream (121))
Discussion: the overall objective here is to "lock down" this script
that serves as a producer for the producer-consumer pair to be used
in the `sync` script. this is an attempt to solidify our new
architecture through abstraction - the khnong cluster itself was not
the direct target of this work, but we stand to gain by putting it
in the scope, from both sides (that is, both for the sake of the
syncing script and for the sake of the khong cluster).

we had dangling diffs for the khong scripts from our "rumskalla" that
we last worked on (gulp) maybe several weeks ago. in the ensuing time
between when we first spiked this datasource script (producer) and now,
the landscape changed among several axes:

  - the runtime context pattern continues to gain precedence over
    the stream lingua franca

  - we learned to appreciate in deeper detail the actual structure
    of the one khong page

  - we discovered that the soupy HTML was much worse than we thought

This commit "locks down" our messy, ad-hoc script that in one way
attempts to protet it from erosion. however the target script (not
tests) will incur us cost when we have to change it because the real
life page structure changes.

(such an incident may or may not lead to an effort at cleaving this
work into something more abstract and resilient. we'll have to cross
that bridge when we get to it.)

Also:
  - add listening thing for debugging
  - add simple emission recorder