(-sakin-agac.18.2 upgrade to python 3.7.0, pipenv i tatsu (206, 1))
Synopsis: this has only 2 sidesystems updated to reflect the upgrade
to python 3.7. we will catch the others up later. This commit should
be the last one for a while that worries about `sakin-agac`. commits
from here forward will be focused on `tag-lyfe` for a while.

Discussion:

upgrading from python 3.6 to 3.7 introduced one category of problems:
this lead to some of our dependency packages issuing warnings; which in
turn caused our runtime to fail because of the overly strict settings
we run python with. rather than loosening the settings, we awfully
manually edited a few lines in the real code of our dependencies
(because so far it has been straightforward) this is awful.

Small points:
  - yikes note that it appears something changed in perhaps `testunit`
    that makes its hacking of the `sys.path` different such that we
    have to change how we hack it in this one place. this will have
    consequences for all the other sub-modules
  - added pycodestyle (used to be called pep8 but guido said don’t)
    so that out of the box from the Pipfile this oft-needed module is
    installed. (compliance to it is something we would make a
    post-commit to if we were so inclined. that is, it’s required.)