(-tag-lyfe.4.4 case study 2: after sync, looks are deceiving..)
ok so now that we did the thing we did in the last commit, our
"sync" operation looks like it yielded some as-expected results:

three items have been added to the collection.

for posterity, we are keeping this tiny commit as-is just to
illustrate this machine-generated edit to our collection. but
in the next commit we'll discuss the problem with it..