(-tag-lyfe.4.3 case study 3: lack of pruning and good keying..)
..as introduced in the previous commit; on first impressions it looks
like the sync yielded a useful outcome: adding three items. but upon
visual inspection we see that some of these are not new items but
rather they are name-changes to existing items.

in this commit (if you look carefully) what we are doing is taking
the taggings that were on the item with the previous name and putting
them "on" the item with the new name, and discarding the item with the
old name.

one way to have safeguarded against this would have been to make our
key-function smarter in a more domain-tailored way: if (in the context
of heroku add-ons) these items are always going to have the same URL
but maybe a different surface label over time, then of course it would
behoove us to use only the URL as the key. (but in other domains like
the previously mentioned parser-generator dim-sum; the simplfying
algorithm that is sort of hurting us here yielded dividends there..)

anyway this dichotomy of there being a different optimal keyer for
the collection domain is food for thought (not something that we saw
coming beforehand).

to make this perhaps even more confusuing; what we actually did is

  1) for each of 3 cached HTML documents from 3 points in time of
     the heroku source page thing (we put these aside at the time to
     do this very research we are sharing now); we generated 3
     different JSON stream dumps of the producer script, using that
     cached file as _the_ cached file. so that is, for the "same"
     webpage at 3 different points in time we took three snapshots
     of it and kept the simplifeid versions of it (the JSON dump)

  2) we made a plain old diff across each of the 2 hops between these
     three files (so, 2 diffs)

for the purposes of illustration, here is the exact contents of the
first of those two diffs:

    52c52
    < {"add_on": "[Einstein
    Vision](https://devcenter.heroku.com/articles/einstein-vision)"}
    ---
    > {"add_on": "[Einstein Vision and
    > Language](https://devcenter.heroku.com/articles/einstein-vision)"}
    86d85
    < {"add_on": "[Kronus](https://devcenter.heroku.com/articles/kronus)"}
    129c128
    < {"add_on": "[Remora
    Backup](https://devcenter.heroku.com/articles/remora-backup)"}
    ---
    > {"add_on": "[Remora Offsite
    > Backup](https://devcenter.heroku.com/articles/remora-backup)"}

quite nicely, this shows exactly what happend in the ensuing time
between those two "scrapes": two items were renamed and one was removed.
(mind you, to recognize the renames as renames you need to know the
priors (and also make some human inductive judgement that has some
algorithm we don't explicitly know here).)..

(for completness, here's the second of the two diffs. not as interesting:)
    92a93
    > {"add_on": "[Mailgun Email
    > Validation](https://devcenter.heroku.com/articles/mailgun-validations)"}

the main side-takes (if not uptake) of all this is to say this: for
the synchronization to have worked perfectly here without human
intervention would have required something like this:

  1. a more domain-tailored keyer (certianly possible with what we
     have now, but we'd like to wait until we are doing markdown
     parsing properly instead of just as a hack)

  2. _pruning_ (at least as emission of a notice).

we haven't talked about pruning yet but that's a whole thing..

in this commit we "pruned" manually the one item that for whatever
reason went away in the far collection (this "kronus" thing).