(-tag-lyfe.4.2 case study 4: manual pruning..)
finally (and as segued to nicely in the previous commit) in this
commit we show that we manully pruned _three_ now defunct items.

the way we actually arrived at determining that these items were the
set difference of the near and far constituency sets .. was HORRENDOUS:

we did a line-by-line visual comparison of our markdown file on the
left and the JSON dump on the right. although it only took a few
minutes to spot the culprit, it is absolutely something we should
want to automate, at least as an option.

(as we write this it's occurring to us that the interleaving algorithm
might not play nice with pruning. well, later for that.)