(-tag-lyfe.4.1 finish this phase of dim-sum note gathering)
Ironically or not, when we got to the end-ish of the long road of
“tag life”, we were hit with some limitations of our approach and
(after using a more detailed, manual approach discussed below) we
ended up determining that we want nothing from the set of 177,
probably.

it is not to say that the approach is not useful generally. we like
the data in the format we have it and we love what we are now able to
do with it. also we are happy that this toolkit will help our data
“scale” into the future in that we can always update it in a mostly
automated way.

so a few things happened in this step:

  - not especially surprising, but we discovered that heroku has
    what we’ll call “semi-hidden metadata”: add-ons have one category
    and given any one add-on you can see its category and then get
    to a listing of all the add-ons under that category; but (it
    appears) you cannot simply see a listing of all the categories.
    yeah, we could etc and we are thinking about it..

  - but the point is heroku’s information architecture could have
    gotten us to more or less the same reduced search space we were
    after: a listing of the 5 or so items they categorize as
    “Content Management”.

  - when it came down to actually assessing each item, we
    “ran out of space” in our dim-sum table to hold all the information
    we wanted to hold. (yes it’s certainly possible to do so, but it
    wouldn’t have been practical/aesthetic)

  - the above gave rise to “single file items”; a thing we may do the
    obvious thing with.

anyway, this is just a snapshot we want to publish now for reasons.
(as it stands now github is our publishing platform and we want to
bounce some of our documents off of colleagues.)

(times)
  08-09 07:45  tired now, but begin cursory review of heroku
        08:02  discover issue with ordering
        08:32  captain's log: OK i have semi-automatically updated
               my list of heroku add-ons. now gonna use filter-by to
                look at only CMS's
        09:07  woah got sucked in to prism. the good news is it is
               hosted so we won't be tempted
        10:35  been reading about the heroku static buildpack for a while 
        11:02  zz
  08-10 05:48  now