(-tag-lyfe.3.4 case study step 1: Begin new roadmap)
This commit (the first of four) begins an experimental attempt at
augmenting our usual methedology of digraph-assisted roadmap making.

What lead us to this is that we felt like we were staring down the
barrel of complete paralysis. we had a hazy idea of where we wanted
to get to and a bunch of .. "states" we thought we wanted to go thru,
but weren't sure of how to procede and (most significantly) of how
to line up the states in some order and frame them in terms of what
actions to take.

The first big departure here in terms of methodology was that this
"states" thing was new in the first place (to us; not to the world
of theory. this is intentionally supposed to sort of resemble a
discrete fininite state automata, a theoretical system that's
exhaustively understood if not by us then by the broader community.).

the other big departures here we will introduce in the following
commits as they are illustrated by the changesets.

(times - we stashed this and were never sure when to bring it back)
  06-16 05:20  begin human indexing every heroku add-on
        07:36  done
        08:16  discovered perch CMS
        21:58  new five step
  06-17 04:53  we have open sync request. now gonna write the CLI
               so we can guide the testing

(the below times encompass the following for case study snapshots)
(times)
  08-10 05:51  begin reading in-depth about jekyll on heroku
        06:54  finished long, fruitful conv with pokey about heroku
               et al. now begin reading 1 of 2 blogs about etc
        06:58  uh-oh what is hugo
        08:45  (just checking in with the document - we are looking
               at the hugo buildpack)
        08:51  tentatively we're gonna assume that hugo will 'just work'
               on heroku and so we'll read a little of its own
               documentation
        02:57  ok just spent about 15 minutes reading the how to make
               a good commit message doohah
        05:36  ok gonna invest in hugo for a bit. (why?) gonna make a
               dependency graph for it..
        08:54  finished thing that will be step four. zz
  08-11 13:27  begin, only half working (baby jon)
        16:11  now